15: Support writing facts.yaml
This will now write facts in confdir/facts.yaml on both platforms, a
script is included to generate the facts.yaml server.cfg is configured
accordingly.

A Cron job or a Windows Task Scheduler job is added to keep this
updated, untested on Windows.

Additionally adds the filemgr agent and prep for release 0.0.7