(#47) Run fact refresher only when needed
Previously the facts refresher would run during every Puppet run, nice
to have fresh facts but can become annoying.

This updates the code to use creates to check if the facts file exist
already and do not run the script.

When facts_refresh_interval = 0 the script is always run by Puppet since
there will be no cron to do it