(#69) adjust dependencies for librarian
This module has a bunch of choria dependencies on it which I put there
so people get a nice simple install process - just install the
mcollective module and choria comes with.

But in turn the choria modules need this module as a dependency so they
can use the plugin installer.

puppet module install does not complain about this circular dependency
but librarian does so this makes the mcollective module be standalone
and the choria one will bring all its dependencies which is probably the
more correct design anyway