(#74) Fix some paths on windows
The libdir was incorrectly configured using the wrong hiera key, but it
also had the wrong value.  Additionally a Puppet bug PUP-7061 is now
somehow breaking lookup() so have to duplicate this path for now

The fact is especially important since without it PQL discovery would
not work, and the way it is is a bit dodgy with hard coded paths - but
we cant really do much better for the fact I suspect since I cant use
a template to generate it.  Might have to think of a better way to get
sub collectives in the future