(#115) give top level properties priority over some settings
There's this unfortunate mix of things - libdir, logdir etc set at the
class and a mashup of settings from hashes.

Hashes are fine to some extend but for things that must be shared across
modules - like libdir being used by every module plugin to install
themselves - we need this as a top level var.

Ideally other settings too would be elevated but there's the ever
present hell of client vs server config which is unavoidable due to how
mcollective configs are designed

Worse the client/server configs cannot be enumerated because every
plugin ever can have any amount of variables in there.

So this solves the basic problem for libdir by making using its top
level value to override the data hashes giving the top setting priority
in the hash merges

This will always be nasty as long as mco config is as it is :(