(#163) Workaround for YAML aliasing
With Puppet 5.5, the refresh_facts.rb produce a YAML file containing
aliases, which are not properly processed (choria-io/mcollective-choria#455).

This workaround prevents the same object to be assigned to
mcollective::client::libdir and mcollective::server::libdir by
duplicating them.  The Psych serializer then consider them as two
distinct objects and does not emit an alias.