(#166) create JSON representations of DDL files
This creates a basic JSON representation of a agent DDL files when
packaging the agent.

The intention is that this will be used by the go daemon to figure out
if its even worth dispatching a request to the ruby agent - is it a
valid action lets say - before doing the much slower work of spwning
ruby.