Update metadata.json to allow Puppet 6.x (#227)
This module supports Puppet 6 already, but the metadata does not document this. This is a tiny change to allow 6.x to work

Note: I have explicitly chosen not to mark support as simply `>= 4.9.0` as it's not clear to me that all future Puppet versions will automatically "just work". Having said that `choria/choria` has this so I'm not sure if @ripienaar has a different idea.