Allow specifying install_options for gem_dependencies
Use case:
On sites behind a corporate proxy, this is needed to be able
to specify the proxy address consistently. The gem command will
pick up the environment `http_proxy` and `no_proxy` environment
variables if set, but when running under sudo, cron or system service
these variables may not be available.