(#250) Avoid YAML aliases
If Facter build multiple facts from the same object, the serialization
to YAML will generate aliases to avoid duplication.

Aliases can be abused and can be used to cause DOS, so it is advised to
use YANL.safe_load when loading YAML which does not support aliases by
default.  It is done by choria-io/mcorpc-ruby-support.

As a consequence, when a single object is used by more than one fact,
the generated "generated-facts.yaml" file cannot be load.

This commit convert the facts hash to JSON, parse the JSON back to a
new Hash (where all references to the same object will be lost, so
arrays with the same content or different objects) and then converted to
YAML.